Renamed and moved GraphCommonsGenerators to more appropriate place
